+++ modules/browsiRtdProvider.js
@@ -13,22 +13,19 @@
  * @property {string} pubKey
  * @property {string} url
  * @property {?string} keyName
- * @property {number} auctionDelay
  */
 
 import {config} from '../src/config.js';
 import * as utils from '../src/utils';
 import {submodule} from '../src/hook';
-import {ajax} from '../src/ajax';
 
 /** @type {string} */
 const MODULE_NAME = 'realTimeData';
 /** @type {ModuleParams} */
 let _moduleParams = {};
-/** @type {null|Object} */
-let _data = null;
-/** @type {null | function} */
-let _dataReadyCallback = null;
+
+export let _resolvePromise = null;
+const _waitForData = new Promise(resolve => _resolvePromise = resolve);
 
 /**
  * add browsi script to page
@@ -39,8 +36,6 @@ export function addBrowsiTag(bptUrl) {
   script.async = true;
   script.setAttribute('data-sitekey', _moduleParams.siteKey);
   script.setAttribute('data-pubkey', _moduleParams.pubKey);
-  script.setAttribute('prebidbpt', 'true');
-  script.setAttribute('id', 'browsi-tag');
   script.setAttribute('src', bptUrl);
   document.head.appendChild(script);
   return script;
@@ -53,9 +48,9 @@ export function addBrowsiTag(bptUrl) {
 function collectData() {
   const win = window.top;
   const doc = win.document;
-  let browsiData = null;
+  let historicalData = null;
   try {
-    browsiData = utils.getDataFromLocalStorage('__brtd');
+    historicalData = JSON.parse(utils.getDataFromLocalStorage('__brtd'))
   } catch (e) {
     utils.logError('unable to parse __brtd');
   }
@@ -65,56 +60,34 @@ function collectData() {
       sk: _moduleParams.siteKey,
       sw: (win.screen && win.screen.width) || -1,
       sh: (win.screen && win.screen.height) || -1,
-      url: encodeURIComponent(`${doc.location.protocol}//${doc.location.host}${doc.location.pathname}`),
+      url: encodeURIComponent(`${doc.location.protocol}//${doc.location.host}${doc.location.pathname}`)
     },
-    ...(browsiData ? {us: browsiData} : {us: '{}'}),
+    ...(historicalData && historicalData.pi ? {pi: historicalData.pi} : {}),
+    ...(historicalData && historicalData.pv ? {pv: historicalData.pv} : {}),
     ...(document.referrer ? {r: document.referrer} : {}),
     ...(document.title ? {at: document.title} : {})
   };
-  getPredictionsFromServer(`//${_moduleParams.url}/prebid?${toUrlParams(predictorData)}`);
-}
-
-export function setData(data) {
-  _data = data;
-
-  if (typeof _dataReadyCallback === 'function') {
-    _dataReadyCallback(_data);
-    _dataReadyCallback = null;
-  }
-}
-
-/**
- * wait for data from server
- * call callback when data is ready
- * @param {function} callback
- */
-function waitForData(callback) {
-  if (_data) {
-    _dataReadyCallback = null;
-    callback(_data);
-  } else {
-    _dataReadyCallback = callback;
-  }
+  getPredictionsFromServer(`//${_moduleParams.url}/bpt?${serialize(predictorData)}`);
 }
 
 /**
  * filter server data according to adUnits received
- * call callback (onDone) when data is ready
  * @param {adUnit[]} adUnits
- * @param {function} onDone callback function
+ * @return {Object} filtered data
+ * @type {(function(adUnit[]): Promise<(adUnit | {}) | never | {}>)}}
  */
-function sendDataToModule(adUnits, onDone) {
-  try {
-    waitForData(_predictionsData => {
+function sendDataToModule(adUnits) {
+  return _waitForData
+    .then((_predictionsData) => {
       const _predictions = _predictionsData.p;
       if (!_predictions || !Object.keys(_predictions).length) {
-        return onDone({});
+        return ({})
       }
       const slots = getAllSlots();
       if (!slots) {
-        return onDone({});
+        return ({})
       }
-      let dataToReturn = adUnits.reduce((rp, cau) => {
+      let dataToResolve = adUnits.reduce((rp, cau) => {
         const adUnitCode = cau && cau.code;
         if (!adUnitCode) { return rp }
         const predictionData = _predictions[adUnitCode];
@@ -128,11 +101,11 @@ function sendDataToModule(adUnits, onDone) {
         }
         return rp;
       }, {});
-      return onDone(dataToReturn);
+      return (dataToResolve);
+    })
+    .catch((e) => {
+      return ({});
     });
-  } catch (e) {
-    onDone({});
-  }
 }
 
 /**
@@ -175,44 +148,42 @@ export function isIdMatchingAdUnit(id, allSlots, whitelist) {
  * @param {string} url server url with query params
  */
 function getPredictionsFromServer(url) {
-  ajax(url,
-    {
-      success: function (response, req) {
-        if (req.status === 200) {
-          try {
-            const data = JSON.parse(response);
-            if (data && data.p && data.kn) {
-              setData({p: data.p, kn: data.kn});
-            } else {
-              setData({});
-            }
-            addBrowsiTag(data.u);
-          } catch (err) {
-            utils.logError('unable to parse data');
-            setData({})
-          }
-        } else if (req.status === 204) {
-          // unrecognized site key
-          setData({});
-        }
-      },
-      error: function () {
-        setData({});
-        utils.logError('unable to get prediction data');
+  const xmlhttp = new XMLHttpRequest();
+  xmlhttp.onreadystatechange = function() {
+    if (xmlhttp.readyState === 4 && xmlhttp.status === 200) {
+      try {
+        var data = JSON.parse(xmlhttp.responseText);
+        _resolvePromise({p: data.p, kn: data.kn});
+        addBrowsiTag(data.u);
+      } catch (err) {
+        utils.logError('unable to parse data');
       }
     }
-  );
+  };
+  xmlhttp.onloadend = function() {
+    if (xmlhttp.status === 404) {
+      _resolvePromise({});
+      utils.logError('unable to get prediction data');
+    }
+  };
+  xmlhttp.open('GET', url, true);
+  xmlhttp.onerror = function() { _resolvePromise({}) };
+  xmlhttp.send();
 }
 
 /**
  * serialize object and return query params string
- * @param {Object} data
+ * @param {Object} obj
  * @return {string}
  */
-function toUrlParams(data) {
-  return Object.keys(data)
-    .map(key => key + '=' + encodeURIComponent(data[key]))
-    .join('&');
+function serialize(obj) {
+  let str = [];
+  for (let p in obj) {
+    if (obj.hasOwnProperty(p)) {
+      str.push(encodeURIComponent(p) + '=' + encodeURIComponent(obj[p]));
+    }
+  }
+  return str.join('&');
 }
 
 /** @type {RtdSubmodule} */
@@ -226,7 +197,7 @@ export const browsiSubmodule = {
    * get data and send back to realTimeData module
    * @function
    * @param {adUnit[]} adUnits
-   * @param {function} onDone
+   * @returns {Promise}
    */
   getData: sendDataToModule
 };
@@ -236,7 +207,6 @@ export function init(config) {
     try {
       _moduleParams = realTimeData.dataProviders && realTimeData.dataProviders.filter(
         pr => pr.name && pr.name.toLowerCase() === 'browsi')[0].params;
-      _moduleParams.auctionDelay = realTimeData.auctionDelay;
     } catch (e) {
       _moduleParams = {};
     }
